Button Amimation
My first animation code

+++ To Do List 1/To Do List 1/Views/ListView.swift
@@ -12,22 +12,29 @@ struct ListView: View {
   
     
     var body: some View {
-        List {
-            ForEach( listViewModel.items) {item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item: item)
-                            
-                        }
-                    }
-               
+        ZStack{
+            if listViewModel.items.isEmpty{
+                
+                NoItemsView()
+            }else{
+                List {
+                     ForEach( listViewModel.items) {item in
+                       ListRowView(item: item)
+                                .onTapGesture {
+                                withAnimation(.linear){
+                                            listViewModel.updateItem(item: item)
+                                            
+                                         }
+                                     }
+                               
+                                  }
+                                .onDelete(perform: listViewModel.deleteItem)
+                                .onMove(perform: listViewModel.moveItem)
+                            }
+                    .listStyle(PlainListStyle())
+                   }
             }
-                .onDelete(perform: listViewModel.deleteItem)
-                .onMove(perform: listViewModel.moveItem)
-            }
-            
-        .listStyle(PlainListStyle())
+        
         .navigationTitle("Todo List 📝")
         .navigationBarItems(
                             leading: EditButton(),
